// apps/frontend/packages/api/base/codegen/types/CorrectionStatus2.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type CorrectionStatus2 = ("DRAFT" | "UNDER_CFO_REVIEW" | "PARTIALLY_APPROVED" | "RETURNED_FOR_REVISION" | "RESUBMITTED" | "ALL_CFO_APPROVED" | "SENT_TO_DTOE" | "UNDER_DTOE_REVIEW" | "RETURNED_BY_DTOE" | "APPROVED_BY_DTOE");